Rename the base namespace to "ruhrpottmetaller"
