+++ deploy/composer.json
@@ -1,7 +1,7 @@
 {
     "autoload": {
         "psr-4": {
-            "rpmetaller\\": "rpmetaller"
+            "ruhrpottmetaller\\": "ruhrpottmetaller"
         }
     }
 }
