+++ deploy/ruhrpottmetaller/Controller.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace rpmetaller;
+namespace ruhrpottmetaller;
 
 //The controller
 class Controller
