+++ deploy/ruhrpottmetaller/ModelVenue.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace rpmetaller;
+namespace ruhrpottmetaller;
 
 /**
  * Class to acces and maintain venue data.
