+++ deploy/ruhrpottmetaller/View.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace rpmetaller;
+namespace ruhrpottmetaller;
 
 /**
  * View class which is in charge of displaying the data.
