+++ deploy/web/index.php
@@ -8,6 +8,6 @@ include '../vendor/autoload.php';
 // Merge $_GET und $_POST
 $request = array_merge($_GET, $_POST);
 // Create controller object
-$Controller = new rpmetaller\Controller($request);
+$Controller = new ruhrpottmetaller\Controller($request);
 // Display the output of web application.
 echo $Controller->getOutput();
